xds/federation: fix percent encoding on server side
