+++ xds/src/main/java/io/grpc/xds/XdsServerWrapper.java
@@ -24,7 +24,6 @@ import com.google.auto.value.AutoValue;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.net.UrlEscapers;
 import com.google.common.util.concurrent.SettableFuture;
 import io.grpc.Attributes;
 import io.grpc.InternalServerInterceptors;
@@ -196,7 +195,7 @@ final class XdsServerWrapper extends Server {
     }
     String replacement = listenerAddress;
     if (listenerTemplate.startsWith(XDSTP_SCHEME)) {
-      replacement = UrlEscapers.urlFragmentEscaper().escape(replacement);
+      replacement = XdsClient.percentEncodePath(replacement);
     }
     discoveryState = new DiscoveryState(listenerTemplate.replaceAll("%s", replacement));
   }
